refactor: Team 컴포넌트 분리
- min27604/baseball/#10
- Styled component였던 Team 컴포넌트를 일반 컴포넌트로 분리
