+++ FE/react-baseball/src/components/Game.jsx
@@ -1,25 +1,16 @@
-import { useHistory } from 'react-router';
 import styled from 'styled-components';
+import Team from './Team';
 
 const Game = ({ game }) => {
-  const history = useHistory();
-  const handleClickTeam = () => history.push('/play-screen');
-  // 클릭하면 GET /games/{gameId} 요청 보내서
-  // 입장 가능한 게임인지 데이터를 받아오고, 상태를 set한다.
-
   return (
     <GameLi>
       <Wrapper>
         <MatchNumber>GAME {game.id}</MatchNumber>
       </Wrapper>
       <Wrapper>
-        <Team type='home' onClick={handleClickTeam}>
-          {game.home_team.name}
-        </Team>
+        <Team type='home' game={game} />
         <Versus>vs</Versus>
-        <Team type='away' onClick={handleClickTeam}>
-          {game.away_team.name}
-        </Team>
+        <Team type='away' game={game} />
       </Wrapper>
     </GameLi>
   );
@@ -44,20 +35,6 @@ const Wrapper = styled.div`
 const MatchNumber = styled.span`
   color: red;
   margin-bottom: 6px;
-`;
-
-const Team = styled.div`
-  /* font-size: ${({ theme }) => theme.fontSizes.XL}; */
-  font-size: 32px;
-  font-weight: bold;
-  color: ${({ theme }) => theme.colors.darkGray};
-  width: 180px;
-  text-align: ${({ type }) => (type === 'home' ? 'right' : 'left')};
-
-  &:hover {
-    cursor: pointer;
-    color: ${({ theme }) => theme.colors.green};
-  }
 `;
 
 const Versus = styled.div`
